Refactored Codebase and added term save/load
